NFDiv-3789 Refactor class name for consistency
