+++ src/test/java/uk/gov/hmcts/divorce/systemupdate/service/task/GenerateApplyForFinalOrderDocumentTest.java
@@ -36,8 +36,8 @@ import static uk.gov.hmcts.divorce.notification.CommonContent.IS_JOINT;
 import static uk.gov.hmcts.divorce.notification.CommonContent.PARTNER;
 import static uk.gov.hmcts.divorce.notification.FormatUtil.DATE_TIME_FORMATTER;
 import static uk.gov.hmcts.divorce.notification.FormatUtil.formatId;
-import static uk.gov.hmcts.divorce.systemupdate.service.task.GenerateApplyForConditionalOrderDocument.FIRST_NAME;
-import static uk.gov.hmcts.divorce.systemupdate.service.task.GenerateApplyForConditionalOrderDocument.LAST_NAME;
+import static uk.gov.hmcts.divorce.systemupdate.service.task.ApplyForConditionalOrderTemplateContent.FIRST_NAME;
+import static uk.gov.hmcts.divorce.systemupdate.service.task.ApplyForConditionalOrderTemplateContent.LAST_NAME;
 import static uk.gov.hmcts.divorce.testutil.ClockTestUtil.getExpectedLocalDate;
 import static uk.gov.hmcts.divorce.testutil.ClockTestUtil.setMockClock;
 import static uk.gov.hmcts.divorce.testutil.TestConstants.TEST_CASE_ID;
